+++ src/Controller/ReservationController.php
@@ -2,12 +2,11 @@
 
 namespace App\Controller;
 
-use App\Entity\Commercial;
 use App\Entity\Reservation;
 use App\Form\ReservationType;
-use App\Service\CommercialApiService;
 use App\Service\ReservationApiService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -16,16 +15,21 @@ use Symfony\Component\Routing\Annotation\Route;
 class ReservationController extends AbstractController
 {
     #[Route('/', name: 'app_reservation_index', methods: ['GET'])]
-    public function index(ReservationApiService $ReservationApiService): Response
+    public function index(ReservationApiService $ReservationApiService, Security $security): Response
     {
+        if ($this->isGranted("ROLE_ADMIN")) $data = $ReservationApiService->getReservations();
+        else $data = $ReservationApiService->getReservations(['id_user' => $this->getUser()->getId()]);
         return $this->render('reservation/index.html.twig', [
-            'reservations' => $ReservationApiService->getReservations(),
+            'reservations' => $data,
         ]);
     }
 
     #[Route('/{id}', name: 'app_reservation_show', methods: ['GET'])]
     public function show(Reservation $reservation): Response
     {
+        if ($reservation->getIdUser()->getId() != $this->getUser()->getId()  && !$this->isGranted("ROLE_ADMIN")) {
+            return $this->redirectToRoute('app_reservation_index', [], Response::HTTP_SEE_OTHER);
+        }
         return $this->render('reservation/show.html.twig', [
             'reservation' => $reservation,
         ]);
@@ -35,6 +39,9 @@ class ReservationController extends AbstractController
     public function edit(Request $request, Reservation $reservation, ReservationApiService $ReservationApiService): Response
     {
         // dd($reservation);
+        if ($reservation->getIdUser()->getId() != $this->getUser()->getId() && !$this->isGranted("ROLE_ADMIN")) {
+            return $this->redirectToRoute('app_reservation_index', [], Response::HTTP_SEE_OTHER);
+        }
         $form = $this->createForm(ReservationType::class, $reservation);
         $form->handleRequest($request);
 
@@ -53,6 +60,9 @@ class ReservationController extends AbstractController
     #[Route('/{id}', name: 'app_reservation_delete', methods: ['POST'])]
     public function delete(Request $request, Reservation $reservation, ReservationApiService $ReservationApiService): Response
     {
+        if ($reservation->getIdUser()->getId() != $this->getUser()->getId() && !$this->isGranted("ROLE_ADMIN")) {
+            return $this->redirectToRoute('app_reservation_index', [], Response::HTTP_SEE_OTHER);
+        }
         if ($this->isCsrfTokenValid('delete'.$reservation->getId(), $request->request->get('_token'))) {
             $ReservationApiService->DeleteReservation($reservation->getId());
         }
