+++ src/Controller/VoyageurController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Voyageur;
 use App\Form\VoyageurType;
+use App\Service\ReservationApiService;
 use App\Service\VoyageurApiService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -14,8 +15,19 @@ use Symfony\Component\Routing\Annotation\Route;
 class VoyageurController extends AbstractController
 {
     #[Route('/', name: 'app_voyageur_index', methods: ['GET'])]
-    public function index(VoyageurApiService $VoyageurApiService, array $_route_params): Response
+    public function index(VoyageurApiService $VoyageurApiService, array $_route_params, ReservationApiService $reservationApiService): Response
     {
+        if($this->isGranted('ROLE_ADMIN')){
+            $reservation =  $reservationApiService->getReservation($_route_params['idR']);
+            if ($reservation->getId() == null ) {
+                return $this->redirectToRoute('app_reservation_index', [], Response::HTTP_SEE_OTHER);
+            }
+            else{
+                if ($reservation->getIdUser()->getId() != $this->getUser()->getId()) {
+                    return $this->redirectToRoute('app_reservation_index', [], Response::HTTP_SEE_OTHER);
+                }
+            }
+        }
         return $this->render('voyageur/index.html.twig', [
             'voyageurs' => $VoyageurApiService->getVoyageurs($_route_params['idR']),
             'idR' => $_route_params['idR'],
@@ -23,9 +35,15 @@ class VoyageurController extends AbstractController
     }
 
     #[Route('/new', name: 'app_voyageur_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, VoyageurApiService $VoyageurApiService, array $_route_params): Response
+    public function new(Request $request, VoyageurApiService $VoyageurApiService, array $_route_params, ReservationApiService $reservationApiService): Response
     {
         $voyageur = new Voyageur();
+        if($this->isGranted('ROLE_ADMIN')){
+            $reservation =  $reservationApiService->getReservation($_route_params['idR']);
+            if ($reservation->getIdUser()->getId() != $this->getUser()->getId() && !$this->isGranted("ROLE_ADMIN")) {
+                return $this->redirectToRoute('app_reservation_index', [], Response::HTTP_SEE_OTHER);
+            }
+        }
         $form = $this->createForm(VoyageurType::class, $voyageur);
         $form->handleRequest($request);
 
@@ -45,6 +63,9 @@ class VoyageurController extends AbstractController
     #[Route('/{id}', name: 'app_voyageur_show', methods: ['GET'])]
     public function show(Voyageur $voyageur, array $_route_params): Response
     {
+        if ($voyageur->getIdReservation()->getIdUser()->getId() != $this->getUser()->getId() && !$this->isGranted("ROLE_ADMIN") ) {
+            return $this->redirectToRoute('app_reservation_index', [], Response::HTTP_SEE_OTHER);
+        }
         return $this->render('voyageur/show.html.twig', [
             'voyageur' => $voyageur,
             'idR' => $_route_params['idR'],
@@ -54,6 +75,9 @@ class VoyageurController extends AbstractController
     #[Route('/{id}/edit', name: 'app_voyageur_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Voyageur $voyageur, VoyageurApiService $VoyageurApiService, array $_route_params): Response
     {
+        if ($voyageur->getIdReservation()->getIdUser()->getId() != $this->getUser()->getId()  && !$this->isGranted("ROLE_ADMIN")) {
+            return $this->redirectToRoute('app_reservation_index', [], Response::HTTP_SEE_OTHER);
+        }
         $form = $this->createForm(VoyageurType::class, $voyageur);
         $form->handleRequest($request);
 
@@ -73,6 +97,9 @@ class VoyageurController extends AbstractController
     #[Route('/{id}', name: 'app_voyageur_delete', methods: ['POST'])]
     public function delete(Request $request, Voyageur $voyageur, VoyageurApiService $VoyageurApiService, array $_route_params): Response
     {
+        if ($voyageur->getIdReservation()->getIdUser()->getId() != $this->getUser()->getId() && !$this->isGranted("ROLE_ADMIN")) {
+            return $this->redirectToRoute('app_reservation_index', [], Response::HTTP_SEE_OTHER);
+        }
         if ($this->isCsrfTokenValid('delete'.$voyageur->getId(), $request->request->get('_token'))) {
             $VoyageurApiService->DeleteVoyageur($voyageur->getId());
         }
